pyproject.toml
[tool.poetry]
name = "python3-genai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-core = "^0.2.33"
langchain-openai = "^0.1.22"
langchain-google-genai = "^1.0.10"
langchain-huggingface = "^0.0.3"
python-dotenv = "^1.0.1"
langchain = "^0.2.15"
prompts = "^0.0.1"
schema = "^0.7.7"
community = "^1.0.0b1"
langchain-community = "^0.2.15"
langchainhub = "^0.1.21"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
